% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21089.R
\name{GradientDescent}
\alias{GradientDescent}
\title{Gradient Descent}
\usage{
GradientDescent(
  x,
  y,
  error,
  maxiter,
  stepmethod = T,
  step = 0.001,
  alpha = 0.25,
  beta = 0.8
)
}
\arguments{
\item{x}{a matrix}

\item{y}{Response variable}

\item{error}{Termination condition, the range of two adjacent search results}

\item{maxiter}{The maximum number of iterations}

\item{stepmethod}{If True,function uses backward descent method to find the step length,else uses fixed step}

\item{step}{Fixed step}

\item{alpha}{Parameter of the retrospective descent method}

\item{beta}{Parameter of the retrospective descent method}
}
\value{
coefficient,iter and error of gradient descent functions
}
\description{
Gradient Descent
}
\examples{
\dontrun{
x<-matrix(iris[1:50,1],50,1)
y<-matrix(iris[1:50,2],50,1)
GradientDescent(x,y,1e-14,1e4,stepmethod=T,step=0.001,alpha=0.25,beta=0.8)
GradientDescent(x,y,1e-14,1e4,stepmethod=F,step=0.001,alpha=0.25,beta=0.8)
}
}
